Fix the imports so that pyega3.py works on the command line
