+++ pyega3/libs/credentials.py
