+++ pyega3/libs/utils.py
