+++ test/test_download_dataset.py
@@ -3,11 +3,11 @@ from unittest import mock
 
 import pytest
 
-from pyega3.data_file import DataFile
-from pyega3.data_set import DataSet
+from pyega3.libs.data_file import DataFile
+from pyega3.libs.data_set import DataSet
 
 
-@mock.patch("pyega3.data_file.DataFile")
+@mock.patch("pyega3.libs.data_file.DataFile")
 def test_calls_download_for_every_file_in_dataset(mocked_datafile, mock_data_server, dataset_with_files,
                                                   mock_server_config,
                                                   mock_auth_client, mock_data_client):
@@ -38,7 +38,7 @@ def test_calls_download_for_every_file_in_dataset(mocked_datafile, mock_data_ser
                                              (num_connections, None, None, 5, 5, DataFile.DEFAULT_SLICE_SIZE))
 
 
-@mock.patch("pyega3.data_file.DataFile")
+@mock.patch("pyega3.libs.data_file.DataFile")
 def test_only_download_available_files(mocked_datafile, mock_server_config, mock_data_server, dataset_with_files,
                                        mock_auth_client, mock_data_client):
     num_connections = 5
@@ -80,14 +80,14 @@ def test_no_error_if_md5_mismatch(mock_server_config, mock_data_server, dataset_
                                   mock_data_client):
     def dfr_throws(p1, p2, p3, p4, p5, p6): raise Exception("bad MD5")
 
-    with mock.patch("pyega3.data_file.DataFile.download_file_retry", dfr_throws):
+    with mock.patch("pyega3.libs.data_file.DataFile.download_file_retry", dfr_throws):
         dataset = DataSet(mock_data_client, dataset_with_files.id)
         dataset.download(1, None, None, 5, 5)
 
 
 def test_download_unknown_dataset_does_not_call_download_file_retry(mock_server_config, mock_data_server,
                                                                     mock_auth_client, mock_data_client):
-    with mock.patch("pyega3.data_file.DataFile.download_file_retry") as mocked_dfr:
+    with mock.patch("pyega3.libs.data_file.DataFile.download_file_retry") as mocked_dfr:
         bad_dataset = "EGAD00000000666"
         dataset = DataSet(mock_data_client, bad_dataset)
         dataset.download(1, None, None, 5, 5)
