+++ test/test_openid.py
@@ -5,8 +5,8 @@ import pytest
 import responses
 
 import test.conftest as common
-from pyega3.auth_client import AuthClient
-from pyega3.credentials import Credentials
+from pyega3.libs.auth_client import AuthClient
+from pyega3.libs.credentials import Credentials
 
 
 @pytest.fixture
